+++ api/TeamLunch/Queries/GetActiveFineRequests.cs
@@ -5,7 +5,7 @@ namespace TeamLunch.Queries;
 
 public static class GetActiveFineRequests
 {
-    public record Query(string userId) : IRequest<List<Response>>;
+    public record Query(string userId, int teamId) : IRequest<List<Response>>;
 
     public class Handler : IRequestHandler<Query, List<Response>>
     {
@@ -18,7 +18,13 @@ public static class GetActiveFineRequests
 
         public async Task<List<Response>> Handle(Query request, CancellationToken cancellationToken)
         {
+            var teamUsers = _db.Teams.Where(x => x.Id == request.teamId).Select(x => x.Users).First();
+            var user = _db.Users.Where(x => x.Id == request.userId).First();
+
+            if (!teamUsers.Contains(user)) throw new UnauthorizedAccessException("User does not have access to this team");
+
             var requests = _db.FineRequests
+                .Where(x => x.TeamId == request.teamId)
                 .Where(x => !x.Responses.Where(r => r.UserId == request.userId).Any())
                 .Where(x => !((x.Finee == request.userId) || (x.Finer == request.userId)))
                 .Select(x => new Response(